+++ src/app/products/services/products.service.ts
@@ -3,13 +3,21 @@ import { ProductModel } from '../models/product.model';
 
 import { productsData } from '../../shared/mocks';
 
+const productsList = Promise.resolve(productsData.map(item => new ProductModel(item)));
+
 @Injectable()
 export class ProductsService {
-  products: ProductModel[] = productsData;
 
   constructor() { }
 
-  getProducts(): ProductModel[] {
-    return this.products;
+  getProducts(): Promise<ProductModel[]> {
+    return productsList;
   }
+
+  getProduct(id: number | string): Promise<ProductModel> {
+    return this.getProducts()
+      .then(products => products.find(product => product.id === +id))
+      .catch(() => Promise.reject('Error in getProduct method'));
+  }
+
 }